3rd Internship Task - E commerce website
Excited to share my latest project as part of my UI/UX Design Internship at CodSoft! 🎉 This task involved designing an e-commerce website that combines aesthetics with functionality to provide a seamless shopping experience. It was an amazing experience to bring this concept to life and understand the nuances of effective e-commerce design!